moved LitElement stories from poly-look to dev-utils/storybook
