+++ dev-utils/storybook/.storybook/lit/main.js
@@ -1,0 +1,3 @@
+module.exports = {
+    stories: ["../../stories/lit-components/*.stories.js"],
+};
