--- feature-utils/poly-look/.storybook/main.js
@@ -1,3 +1,0 @@
-module.exports = {
-  stories: ["../stories/lit-components/*.stories.js"],
-};
