+++ feature-utils/poly-look/package.json
@@ -11,7 +11,6 @@
     "format": "eslint --ext .js,.html . --fix --ignore-path .gitignore && prettier \"**/*.js\" --write --ignore-path .gitignore",
     "test": "jest --coverage test/react-components/* && web-test-runner --coverage",
     "test:watch": "web-test-runner --watch",
-    "storybook": "web-dev-server -c .storybook/server.mjs",
     "build": "rollup -c rollup.config.js",
     "build-downstream": "../../build.js --start @polypoly-eu/poly-look",
     "doc": "jsdoc -R README.md -r src -d ./docs"
@@ -30,8 +29,6 @@
     "@open-wc/testing": "^3.0.3",
     "@testing-library/jest-dom": "^5.16.2",
     "@testing-library/react": "^12.1.3",
-    "@web/dev-server-rollup": "^0.3.13",
-    "@web/dev-server-storybook": "^0.0.2",
     "@web/test-runner": "^0.13.22",
     "@web/test-runner-chrome": "^0.10.5",
     "identity-obj-proxy": "^3.0.0"
